+++ src/tests/legacy/proxies.test.ts
@@ -16,11 +16,11 @@ describe('Value proxies', () => {
 
 		const proxy = booleanProxy(form, 'bool');
 
-		expect(get(form).bool).toStrictEqual(false);
+		expect(get(form).bool).toBe(false);
 
 		proxy.set('true');
 
-		expect(get(form).bool).toStrictEqual(true);
+		expect(get(form).bool).toBe(true);
 	});
 
 	test('intProxy', async () => {
@@ -33,11 +33,11 @@ describe('Value proxies', () => {
 
 		const proxy = intProxy(form, 'int');
 
-		expect(get(form).int).toStrictEqual(0);
+		expect(get(form).int).toBe(0);
 
 		proxy.set('123');
 
-		expect(get(form).int).toStrictEqual(123);
+		expect(get(form).int).toBe(123);
 	});
 
 	test('numberProxy', async () => {
@@ -50,11 +50,27 @@ describe('Value proxies', () => {
 
 		const proxy = numberProxy(form, 'number');
 
-		expect(get(form).number).toStrictEqual(0);
+		expect(get(form).number).toBe(0);
 
 		proxy.set('123.5');
+		expect(get(form).number).toBe(123.5);
 
-		expect(get(form).number).toStrictEqual(123.5);
+		proxy.set('');
+		expect(get(form).number).toBe(NaN);
+	});
+
+	test('numberProxy with empty settings', async () => {
+		const schema = z.object({
+			number: z.number()
+		});
+
+		const superForm = await superValidate(zod(schema));
+		const form = writable(superForm.data);
+
+		const proxy = numberProxy(form, 'number', { zeroIfEmpty: true });
+
+		proxy.set('');
+		expect(get(form).number).toBe(0);
 	});
 
 	test('dateProxy', async () => {
@@ -88,9 +104,9 @@ describe('Field proxies', () => {
 
 		const proxy = fieldProxy(form, 'test[2]');
 
-		expect(get(proxy)).toEqual(2);
+		expect(get(proxy)).toBe(2);
 		proxy.set(123);
-		expect(get(proxy)).toEqual(123);
-		expect(get(form).test[2]).toEqual(123);
+		expect(get(proxy)).toBe(123);
+		expect(get(form).test[2]).toBe(123);
 	});
 });